// src/parser/model/aliases.ts
export const enum ComponentOuterAliasKind {
    /// The alias is to an outer core module.
    CoreModule,
    /// The alias is to an outer core type.
    CoreType,
    /// The alias is to an outer type.
    Type,
    /// The alias is to an outer component.
    Component,
}

/// Represents an alias in a WebAssembly component.
export type ComponentAlias =
    | ComponentAliasInstanceExport
    | ComponentAliasCoreInstanceExport
    | ComponentAliasOuter

export type ComponentFunction =
    | CanonicalFunctionLift
    | ComponentAliasInstanceExport
    | ComponentImport

/// The alias is to an export of a component instance.
export type ComponentAliasInstanceExport = IndexedElement & {
    tag: ModelTag.ComponentAliasInstanceExport,
    /// The alias kind.
    kind: ComponentExternalKind,
    /// The instance index.
    instance_index: ComponentInstanceIndex,
    /// The export name.
    name: string,
}